correção de bug

+++ template-jpa/src/main/java/br/edu/ifpb/es/daw/dao/impl/EmpregadoDAOImpl.java
@@ -4,7 +4,7 @@ import br.edu.ifpb.es.daw.dao.EmpregadoDAO;
 import br.edu.ifpb.es.daw.entities.Empregado;
 import jakarta.persistence.EntityManagerFactory;
 
-public class EmpregadoDAOImpl extends AbstractDAOImpl<Empregado, Integer> implements EmpregadoDAO {
+public class EmpregadoDAOImpl extends AbstractDAOImpl<Empregado, Long> implements EmpregadoDAO {
     
     public EmpregadoDAOImpl(EntityManagerFactory emf) {
             super(Empregado.class, emf);
